Crate hermod

Rapid and seamless customer interaction. Designed by Russell Weas, Jonathan Pence, Do Thien An Duong, Simon Jeon for CS 495 at The University of Alabama.

This wiki only contains documentation for publicly exposed structs/methods in each module. For more information about auto-generated Rust documentation, visit the rustdoc documentation.

All code in this crate is licensed by the file in the GitHub repo.

Modules

auth

Contains methods used for user authentication and authorization.

configuration

Contains structs and helpers related to server configuration.

db

Contains structs that model database tables.

handlers

Contains HTTP Handlers that directly receive and respond to requests to the server.

startup

Contains code neccessary to bootstrap the application and run the server.

telemetry

Contains functions that configure and initialize tracing telemetry, as well as additional helpers.